requirements.txt
requests 
python-dotenv 
pandas 
scikit-learn 
numpy 
SQLAlchemy 
openmeteo-requests 
requests-cache 
retry-requests 
psycopg2
